+++ src/app/api/v1/orders/[id]/route.ts
@@ -4,30 +4,72 @@ import { getOrderById } from "../controllers/getOrderById"
 import { updateOrder } from "../controllers/updateOrder"
 import dbConnect from "@/app/api/mongoose/mongoose"
 import { deleteOrder } from "../controllers/deleteOrder"
+import { checkLogger } from "@/app/api/auth/[...nextauth]/checkLogger"
+import { NextResponse } from "next/server"
+
+
+
+
+
 
 export async function PATCH(_request: Request, { params }: { params: { id: string } }) {
-   await dbConnect()
-   const id = params.id
-   const data = await cloneOrder(id)
-   return Response.json(data, { status: data.code })
+   const user = await checkLogger()
+   if (!user) {
+      return NextResponse.json({ message: 'access denied' }, { status: 401 })
+   }
+   if (user.roll == 'admin') {
+      await dbConnect()
+      const id = params.id
+      const data = await cloneOrder(id)
+      return NextResponse.json(data, { status: data.code })
+   }
+   return NextResponse.json({ message: 'access denied' }, { status: 401 })
 }
+
+
 export async function PUT(request: Request, { params }: { params: { id: string } }) {
-   await dbConnect()
-   const body = await request.json()
-   const id = params.id
-   const data = await updateOrder(id, body)
-   return Response.json(data, { status: data.code })
+
+   const user = await checkLogger()
+   if (!user) {
+      return NextResponse.json({ message: 'access denied' }, { status: 401 })
+   }
+   if (user.roll == 'admin') {
+      await dbConnect()
+      const body = await request.json()
+      const id = params.id
+      const data = await updateOrder(id, body)
+      return NextResponse.json(data, { status: data.code })
+   }
+   return NextResponse.json({ message: 'access denied' }, { status: 401 })
 }
+
+
 export async function GET(request: Request, { params }: { params: { id: string } }) {
-   await dbConnect()
-   const isExpand = getQueryParams(request.url as string, 'expand', false)
-   const id = params.id
-   const data = await getOrderById(id, isExpand)
-   return Response.json(data, { status: data.code })
+   const user = await checkLogger()
+   if (!user) {
+      return NextResponse.json({ message: 'access denied' }, { status: 401 })
+   }
+   if (user.roll == 'admin') {
+      await dbConnect()
+      const isExpand = getQueryParams(request.url as string, 'expand', false)
+      const id = params.id
+      const data = await getOrderById(id, isExpand)
+      return NextResponse.json(data, { status: data.code })
+   }
+   return NextResponse.json({ message: 'access denied' }, { status: 401 })
 }
+
+
 export async function DELETE(request: Request, { params }: { params: { id: string } }) {
-   await dbConnect()
-   const id = params.id
-   const data = await deleteOrder(id)
-   return Response.json(data, { status: data.code })
+   const user = await checkLogger()
+   if (!user) {
+      return NextResponse.json({ message: 'access denied' }, { status: 401 })
+   }
+   if (user.roll == 'admin') {
+      await dbConnect()
+      const id = params.id
+      const data = await deleteOrder(id)
+      return NextResponse.json(data, { status: data.code })
+   }
+   return NextResponse.json({ message: 'access denied' }, { status: 401 })
 }
